frontend/refactor : Move emergency contacts native functions to common.
