+++ Frontend/android-native/mobility-common/src/main/java/in/juspay/mobility/common/MobilityCommonBridge.java
@@ -38,6 +38,7 @@ import android.content.ActivityNotFoundException;
 import android.content.BroadcastReceiver;
 import android.content.ClipData;
 import android.content.ClipboardManager;
+import android.content.ContentResolver;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
@@ -47,6 +48,7 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.content.pm.ResolveInfo;
 import android.content.res.Resources;
+import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.graphics.Canvas;
@@ -73,6 +75,7 @@ import android.os.Handler;
 import android.os.Looper;
 import android.os.VibrationEffect;
 import android.os.Vibrator;
+import android.provider.ContactsContract;
 import android.provider.MediaStore;
 import android.util.Base64;
 import android.util.DisplayMetrics;
@@ -250,6 +253,7 @@ public class MobilityCommonBridge extends HyperBridge {
     protected String CALLBACK = "CALLBACK";
     protected  String NOTIFICATION = "NOTIFICATION";
     protected  String LOG_TAG = MobilityCommonBridge.class.getSimpleName();
+    private static String storeContactsCallBack = null;
     protected String phoneNumber;
     protected String invoice;
     CancellationTokenSource cancellationTokenSource = new CancellationTokenSource();
@@ -4124,9 +4128,90 @@ public class MobilityCommonBridge extends HyperBridge {
                     Toast.makeText(bridgeComponents.getContext(), "Permission Denied", Toast.LENGTH_SHORT).show();
                 }
                 break;
+            case REQUEST_CONTACTS:
+                boolean flag = ContextCompat.checkSelfPermission(bridgeComponents.getContext(), Manifest.permission.READ_CONTACTS) == PackageManager.PERMISSION_GRANTED;
+                String contacts;
+                try {
+                    if (flag) {
+                        contacts = getPhoneContacts();
+                    } else {
+                        JSONArray flagArray = new JSONArray();
+                        contacts = flagArray.toString();
+                    }
+                    contactsStoreCall(contacts);
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+                break;
             default:
                 break;
         }
         return super.onRequestPermissionResult(requestCode, permissions, grantResults);
     }
+
+    //region Store and Trigger CallBack
+    @JavascriptInterface
+    public void contactPermission() {
+        if (ContextCompat.checkSelfPermission(bridgeComponents.getContext(), Manifest.permission.READ_CONTACTS) != PackageManager.PERMISSION_GRANTED) {
+            if (bridgeComponents.getActivity() != null) {
+                ActivityCompat.requestPermissions(bridgeComponents.getActivity(), new String[]{Manifest.permission.READ_CONTACTS}, REQUEST_CONTACTS);
+            }
+        } else {
+            ExecutorManager.runOnBackgroundThread(() -> {
+                try {
+                    contactsStoreCall(getPhoneContacts());
+                } catch (Exception e) {
+                    Log.e(UTILS, "Exception in Contact Permission" + e);
+                }
+            });
+        }
+    }
+
+    public String getPhoneContacts() throws JSONException {
+        ContentResolver contentResolver = bridgeComponents.getContext().getContentResolver();
+        Uri uri = ContactsContract.CommonDataKinds.Phone.CONTENT_URI;
+        JSONArray contacts = new JSONArray();
+        String[] projection =
+                {
+                        ContactsContract.CommonDataKinds.Phone.NUMBER,
+                        ContactsContract.CommonDataKinds.Phone.DISPLAY_NAME
+                };
+        String selection = ContactsContract.Contacts.IN_VISIBLE_GROUP + " = '"
+                + ("1") + "'";
+        String sortOrder = ContactsContract.Contacts.DISPLAY_NAME
+                + " COLLATE LOCALIZED ASC";
+        try (Cursor cursor = contentResolver.query(uri, projection, selection, null, sortOrder)) {
+            if (cursor.getCount() > 0) {
+                while (cursor.moveToNext()) {
+                    String contactNameStr = cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.DISPLAY_NAME));
+                    String contactStr = cursor.getString(cursor.getColumnIndexOrThrow(ContactsContract.CommonDataKinds.Phone.NUMBER));
+                    JSONObject tempPoints = new JSONObject();
+                    tempPoints.put("name", contactNameStr);
+                    tempPoints.put("number", contactStr);
+                    contacts.put(tempPoints);
+                }
+            }
+        }
+        JSONObject flagObject = new JSONObject();
+        flagObject.put("name", "beckn_contacts_flag");
+        flagObject.put("number", "true");
+        contacts.put(flagObject);
+        return contacts.toString();
+    }
+
+    public void contactsStoreCall(String contacts) {
+        if (storeContactsCallBack != null) {
+            String removedDoubleQuotes = contacts.replace("\\\"", "\\\\\"");
+            String removedSingleQuote = removedDoubleQuotes.replace("'", "");
+            String javascript = String.format(Locale.ENGLISH, "window.callUICallback('%s','%s');",
+                    storeContactsCallBack, removedSingleQuote);
+            bridgeComponents.getJsCallback().addJsToWebView(javascript);
+        }
+    }
+
+    @JavascriptInterface
+    public void storeCallBackContacts(String callback) {
+        storeContactsCallBack = callback;
+    }
+
 }
